better my frontend pages and give rest apis final touch

+++ src/main/java/com/Internship/ecommerce_backend_design/Service/ProductService.java
@@ -3,16 +3,20 @@ package com.Internship.ecommerce_backend_design.Service;
 import com.Internship.ecommerce_backend_design.Entity.ProductDetails;
 import com.Internship.ecommerce_backend_design.Entity.Products;
 import com.Internship.ecommerce_backend_design.Repository.ProductRepository;
+import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
     @Autowired
     ProductRepository productRepository;
+    @Autowired
+    private ProductDetailService productDetailService;
 
     public List<Products> getAllCategories(){
         return productRepository.findAll();
@@ -23,6 +27,7 @@ public class ProductService {
         if(byCategory==null){
             byCategory=new Products();
             byCategory.setCategory(category);
+            byCategory.setImage(category+"jpg");
         }
         byCategory.getProductDetailsList().add(productDetails);
         productRepository.save(byCategory);
@@ -30,19 +35,22 @@ public class ProductService {
     public List<ProductDetails> getByCategory(String category){
         Products byCategory = productRepository.findByCategory(category);
         List<ProductDetails> productDetailsList = byCategory.getProductDetailsList();
-        return productDetailsList;
+       return productDetailsList;
     }
-
     @Transactional
-    public void deleteDetail(String name,ProductDetails productDetail){
-        Products product=productRepository.findByCategory(name);
-        product.getProductDetailsList().remove(productDetail);
-        int size=product.getProductDetailsList().size();
-        if(size==0){
-            productRepository.deleteByCategory(product.getCategory());
-            return;
+    public void deleteDetail(ObjectId id){
+        Optional<ProductDetails> detailsbyId = productDetailService.getDetailsbyId(id);
+        if(detailsbyId.isPresent()){
+            ProductDetails productDetails = detailsbyId.get();
+            Products product=productRepository.findByCategory(productDetails.getCategory());
+            product.getProductDetailsList().remove(productDetails);
+            int size=product.getProductDetailsList().size();
+            if(size==0){
+                productRepository.deleteByCategory(product.getCategory());
+                return;
+            }
+            productRepository.save(product);
         }
-        productRepository.save(product);
     }
 
     public void saveCategory(Products product) {
